Move dashboard content to its module
